Add public routes to verify email addresses and resend confirmations for primary email addresses
Allow app_username fields to be passed a string as username that will be kept

When the username as string is known, but no user object has been fetched from the database,
it it easier to just pass this username string to the form than to first fetch a user from the database.
On form submission, this username will then be converted to a user object.

Additionally, it lowers database load and improves page loading speed

Handle case where email to be verified has been removed after the link is sent.

This is the case when a user adds an email address, and removes it immediately.
An activation email will have been sent, but the email address the link refers to
will no longer exist.

It is impossible to show a more specific message containing any information because
it is no longer available in the database. We also do not want to divulge any
identifying information, because the email could have been sent to a third party
if the user made a typo and immediately realised their mistake.

Give public email verification boxes the same size as the login form