Fix issues with saving a new user with properties
Hack to get the user properties out of the persist loop when this entity
itself is persisted.
This is required becaue the user properies have a composite primary key,
which consists of this entity id (unknown until a flush happens) and the property entity id.
Doctrine needs to know the id of this entity before the user properties entities can be
persisted, but it also requires all entities in the mapped collection to be persisted
before, or together with the main entity that is persisted.
The result of this is an unsatisfiable condition, since only persisted entities can
be flushed to the database. That is why the user properties collection gets copied
to a temporary variable before persisting, en gets restored after persisting.