Fixes #6
The _token field is no field on the symfony form, so it can't be detected directly to determine of CSRF
protections are enabled or disabled.
The _token field is always overwritten when the form namespace gets overwritten with array('submit'=>null),
all requests are missing the token.
Since we do not have to detect if CSRF protections are enabled, we can just set the submit button
when it is not present in the request.
This does not hurt security, since web requests missing the _token field are rejected anyways.
API requests are not CSRF protected, so they do not need to send a _token field, only the submit button.