Do not expose whole group graph in /admin/groups and /admin/groups/{name}
Without max depth checks, JMS serializer does not limit the recursion depth
of group members/parents, and a full object graph gets generated causing a huge
amount of queries.

Max depth checks are only enabled when they are specifically requested in the
@View annotation or directly on the SerializerContex in the returned View object,
which was the case in a previous version of the application.

When migrating away from VlRadRestBundle, these properties were forgotten. Tests
did not catch this one because the behavior is explicitly declared undefined in
the API documentation.