Fix LINK & UNLINK API calls for groups and users
GroupController::getAction(), which gets called while resolving
a (UN)LINK request gets passed an incomplete Request, which results
in an exception being thrown when creating the graph form in ::createGraphForm().
More specifically $request->attributes->get('_route_params') returns NULL, which
is an invalid argument for array_merge.
This exception then propagates up the stack until the LinkRequestListener intercepts
it and aborts the resolving of the URL to an object.

The fix is to only call the GroupController::createGraphForm() when there
is a template to be rendered. This can be done by passing a Closure to View::setTemplateData().
The function only gets called when rendering a template, and its return value
is used as additional templating data.