Use better method to override contextual color of button
button_class can be used to use a specific contextual class and remove
the default. Previously, new classes were appended to the default ones,
which only worked due to them being defined after the defaults in the bootstrap
css.
It did not work completely for .btn-link, a blue background (from .btn-primary)
was visible on :active or :focus.